# swagger.yaml
openapi: 3.0.0
info:
  title: API Programming Assignment 5
  version: 1.0.0
  description: REST API for dataset management

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Validates username and password and returns a token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: test
                password:
                  type: string
                  example: 1234
      responses:
        '201':
          description: Login successful – token created
        '401':
          description: Unauthorized – invalid credentials
        '415':
          description: Unsupported Media Type
        '429':
          description: Too many login attempts

  /data:
    get:
      security:
        - TokenAuth: []
      tags:
        - Data
      summary: Get dataset
      description: Returns all data records
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Data'

    post:
      security:
        - TokenAuth: []
      tags:
        - Data
      summary: Create new data
      description: Add new record to dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataInput'
      responses:
        '201':
          description: Created successfully
        '415':
          description: Unsupported Media Type

  /data/{id}:
    delete:
      security:
        - TokenAuth: []
      tags:
        - Data
      summary: Delete data by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Data not found

  /data/search:
    post:
      security:
        - TokenAuth: []
      tags:
        - Data Search
      summary: Search data by forename
      description: Returns matched records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryInput'
      responses:
        '200':
          description: Matching results found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Data'
        '404':
          description: User not found

components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization

  schemas:
    Data:
      type: object
      properties:
        id:
          type: integer
          example: 1
        forename:
          type: string
          example: Roy
        surname:
          type: string
          example: Fielding

    DataInput:
      type: object
      properties:
        forename:
          type: string
          example: Tim
        surname:
          type: string
          example: Berners-Lee

    QueryInput:
      type: object
      properties:
        forename:
          type: string
          example: Roy